Codes for Arduino and Csharp App
Sensor Information:
The  device  consists  of  an  accelerometer  to  capture  voluntary
movements  performed  by  users  who  have  restricted  movement such as
patients with amyotrophic lateral sclerosis.

Firmware Information:
The microcontroller (MCU) receives the analog signal from the
accelerometer, process the data and sends through the usb communication
a message whenever the sensor is triggered. The trigger is activated
when a motion is applied to the sensor.

Software Information:
When a trigger message arrives by USB the application invokes a mouse
event (Click)